+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 1.1.2
 =====
  - Fixed bug in `slice` on `BitVector`/`ByteVector` when passed negative indices
+ - Added an extractor for `ByteVector`, allowing pattern matching on the bytes that make up a vector.
 
 1.1.1
 =====