+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 1.1.2
 =====
  - Fixed bug in `slice` on `BitVector`/`ByteVector` when passed negative indices
- - Added an extractor for `ByteVector`, allowing pattern matching on the bytes that make up a vector.
+ - Added an extractor for `ByteVector`, allowing pattern matching on the bytes that make up a vector
+ - Added an overload of `ByteVector.view` that allows viewing a slice of an array
 
 1.1.1
 =====